requirements.txt
olympuswifi>=0.1.0
Pillow>=9.0.0
numpy>=1.20.0
